% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parperm_boot}
\alias{parperm_boot}
\title{Bootstrapping using parPerm}
\usage{
parperm_boot(x.mat = xtx, y = ymat, columns = 2:3, split = 101,
  num_perms = 100, boots = 10)
}
\arguments{
\item{x.mat}{The desired x matrix that contains all the characteristics data/predictor variables; the matrix should contain the desired column user intends completing the permutation on.}

\item{y}{The desired preprocessed neuroimaging data, with a mask or not.}

\item{columns}{A statement indicating which characteristic/predictor we want analyzed, for multiple categories select the proper columns}

\item{split}{A value that allows user to finely split the processing task to allow for error checking.}

\item{num_perms}{User defined value for number of desired permutations}

\item{boots}{Number of desired bootstraps}
}
\value{
The desired results post-bootstrap
}
\description{
Bootstrapping using parPerm
}
