% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permtest_s.R
\name{run_permtest}
\alias{run_permtest}
\title{Run permutation test using parallel processing}
\usage{
run_permtest(x.mat = xtx, y = ymat, columns = 1, split = 101,
  num_perms = 100)
}
\arguments{
\item{x.mat}{The desired x matrix that contains all the characteristics data/predictor variables; the matrix should contain the desired column user intends completing the permutation on.}

\item{y}{The desired preprocessed neuroimaging data, with a mask or not.}

\item{columns}{A numeric value indicating which characteristic/predictor we want analyzed. User inputs the column number they want permuted.}

\item{split}{A value that allows user to finely split the processing task to allow for error checking.}

\item{num_perms}{User defined value for number of desired permutations}
}
\value{
The desired permutation result
}
\description{
Run permutation test using parallel processing
}
