% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permtest_m.R
\name{run_permtest_multicat}
\alias{run_permtest_multicat}
\title{Run permutation test using parallel processing}
\usage{
run_permtest_multicat(x.mat = xtx, y = ymat, columns = 2:3,
  split = 101, num_perms = 100)
}
\arguments{
\item{x.mat}{the desired x matrix that was preprocessed containing all the characteristics}

\item{y}{the desired preprocessed neuroimaging data}

\item{columns}{which characteristic/predictor we want analyzed, for multiple categories select the columns containing the binary indicators}
}
\value{
the desired permutation result

a <- run_permtest(x.mat = xtx, y = ymat, columns = 2:3, split = 101, num_perms = 100)
}
\description{
Run permutation test using parallel processing
}
