% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_x.R
\name{xmat_trans}
\alias{xmat_trans}
\title{Preprocess x matrix data}
\usage{
xmat_trans(xmat)
}
\arguments{
\item{xmat}{A dataframe containing strictly numerics or binaries, e.g. dataframes created from read_csv or read_excel}
}
\value{
The x-matrix multiplication we desire
}
\description{
Preprocess x matrix data
}
\examples{
\dontrun{
xmat <- read_excel('VMB Variables.xlsx',skip=1)
xtx <- xmat_trans(xmat)
}
}
