% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_y.R
\name{ymat_trans}
\alias{ymat_trans}
\title{Preprocess y matrix data, while double checking all data is nifti}
\usage{
ymat_trans(neurodat)
}
\arguments{
\item{neurodat}{The desired nifti dataset we want analyzed}
}
\value{
A processed matrix ready for permutation test analysis
}
\description{
Preprocess y matrix data, while double checking all data is nifti
}
\examples{
ymat <- ymat_trans(neurodat)
}
