% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_y_mask.R
\name{ymat_trans_mask}
\alias{ymat_trans_mask}
\title{Preprocess y matrix data if user includes a masking file, while double checking all data is nifti}
\usage{
ymat_trans_mask(neurodat, mask)
}
\arguments{
\item{neurodat}{The desired nifti dataset we want analyzed}

\item{mask}{User provided mask}
}
\value{
A processed matrix ready for permutation test analysis

ymat <- ymat_trans(neurodat, mask, mask_logical = TRUE)
}
\description{
Preprocess y matrix data if user includes a masking file, while double checking all data is nifti
}
