% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permtest_s.R
\name{run_permtest}
\alias{run_permtest}
\title{Run permutation test using parallel processing}
\usage{
run_permtest(x.mat = xtx, y = ymat, columns = 1, split = 101,
  num_perms = 100)
}
\arguments{
\item{x.mat}{the desired x matrix that was preprocessed containing all the characteristics}

\item{y}{the desired preprocessed neuroimaging data}

\item{columns}{which characteristic/predictor we want analyzed}
}
\value{
the desired permutation result

a <- run_permtest(x.mat = xtx, y = ymat, columns = 1, split = 101, num_perms = 100)
}
\description{
Run permutation test using parallel processing
}
