% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permtest_m.R
\name{run_permtest_multicat}
\alias{run_permtest_multicat}
\title{Run permutation test using parallel processing}
\usage{
run_permtest_multicat(x.mat = xtx, y = ymat, columns = 2:3,
  split = 101, num_perms = 100)
}
\arguments{
\item{x.mat}{The desired x matrix that contains all the characteristics data/predictor variables; the matrix should contain the desired column user intends completing the permutation on.}

\item{y}{The desired preprocessed neuroimaging data, with a mask or not.}

\item{columns}{A statement indicating which characteristic/predictor we want analyzed, for multiple categories select the proper columns}

\item{split}{A value that allows user to finely split the processing task to allow for error checking.}

\item{num_perms}{User defined value for number of desired permutations}
}
\value{
The desired permutation result

The following three lines of code are necessary for the parallel processing to work, can be changed to user's discretion but this is a known safe method.
###
cores <- detectCores()
cl <- makeCluster(cores[1]-1) #not to overload your computer
registerDoParallel(cl)
###

a <- run_permtest(x.mat = xtx, y = ymat, columns = 1, split = 101, num_perms = 100)
}
\description{
Run permutation test using parallel processing
}
