% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_y.R
\name{ymat_trans}
\alias{ymat_trans}
\title{Preprocess y matrix data, while double checking all data is nifti}
\usage{
ymat_trans(gmvol, mask, mask_logical = FALSE)
}
\arguments{
\item{gmvol}{the desired nifti dataset we want analyzed}

\item{mask}{if we are using a mask because data is too large}

\item{mask_logical}{an indicator whether of not we are using a mask}
}
\value{
a processed matrix ready for permutation test analysis

ymat <- ymat_trans(gmvol, mask, mask_logical = TRUE)
}
\description{
Preprocess y matrix data, while double checking all data is nifti
}
